pyproject.toml
[project]
name = "ax-tool-framework"
version = "0.1.0"
description = "AX Tool Framework v0.1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.0",
  "jsonschema>=4.23.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]